+++ src/werkzeug/exceptions.py
@@ -58,6 +58,7 @@
     :license: BSD-3-Clause
 """
 import sys
+from datetime import datetime
 
 from ._compat import implements_to_string
 from ._compat import integer_types
@@ -592,36 +593,51 @@ class PreconditionRequired(HTTPException):
     )
 
 
-class TooManyRequests(HTTPException):
+class _RetryAfter(HTTPException):
+    """Adds an optional ``retry_after`` parameter which will set the
+    ``Retry-After`` header. May be an :class:`int` number of seconds or
+    a :class:`~datetime.datetime`.
+    """
+
+    def __init__(self, description=None, response=None, retry_after=None):
+        super(_RetryAfter, self).__init__(description, response)
+        self.retry_after = retry_after
+
+    def get_headers(self, environ=None):
+        headers = super(_RetryAfter, self).get_headers(environ)
+
+        if self.retry_after:
+            if isinstance(self.retry_after, datetime):
+                from .http import http_date
+
+                value = http_date(self.retry_after)
+            else:
+                value = str(self.retry_after)
+
+            headers.append(("Retry-After", value))
+
+        return headers
+
+
+class TooManyRequests(_RetryAfter):
     """*429* `Too Many Requests`
 
-    The server is limiting the rate at which this user receives responses, and
-    this request exceeds that rate. (The server may use any convenient method
-    to identify users and their request rates). The server may include a
-    "Retry-After" header to indicate how long the user should wait before
-    retrying.
+    The server is limiting the rate at which this user receives
+    responses, and this request exceeds that rate. (The server may use
+    any convenient method to identify users and their request rates).
+    The server may include a "Retry-After" header to indicate how long
+    the user should wait before retrying.
 
-    .. versionchanged:: 0.16.1
-        ``retry_after_secs`` was added as the first argument, ahead of
-        ``description``.
+    :param retry_after: If given, set the ``Retry-After`` header to this
+        value. May be an :class:`int` number of seconds or a
+        :class:`~datetime.datetime`.
+
+    .. versionchanged:: 1.0
+        Added ``retry_after`` parameter.
     """
 
     code = 429
     description = "This user has exceeded an allotted request count. Try again later."
-
-    def __init__(self, description=None, retry_after_secs=None):
-        """
-        Use the optional value of retry_after_secs to specify the number of seconds
-        to wait for a retry attempt.
-        """
-        HTTPException.__init__(self, description)
-        self.retry_after_secs = retry_after_secs
-
-    def get_headers(self, environ=None):
-        headers = HTTPException.get_headers(self, environ)
-        if self.retry_after_secs:
-            headers.append(("Retry-After", str(self.retry_after_secs)))
-        return headers
 
 
 class RequestHeaderFieldsTooLarge(HTTPException):
@@ -699,10 +715,18 @@ class BadGateway(HTTPException):
     )
 
 
-class ServiceUnavailable(HTTPException):
+class ServiceUnavailable(_RetryAfter):
     """*503* `Service Unavailable`
 
-    Status code you should return if a service is temporarily unavailable.
+    Status code you should return if a service is temporarily
+    unavailable.
+
+    :param retry_after: If given, set the ``Retry-After`` header to this
+        value. May be an :class:`int` number of seconds or a
+        :class:`~datetime.datetime`.
+
+    .. versionchanged:: 1.0
+        Added ``retry_after`` parameter.
     """
 
     code = 503